Move select-methods to own file to delete them
A selected field is better fitting on the value side from the item,
so the user has to implement and handel it if nesseccary.

- removed most of the keybindings from the Update methode,
  except basic movment. To prevent keybinding confusion.
- changed empty Lines() handling to not panic.
- changed Top and Bottom Methods to have error in case of empty list.
- implemented ResetItems to fully replace all items within the list.
- removed UnFocus Methode and changed Focus Method to take a argument.
